modified database schema

+++ src/main/java/com/wifiproject/demo/RSSIController.java
@@ -13,15 +13,13 @@ public class RSSIController {
     @Autowired
     private RSSIMapper rssiMapper;
 
-
-
     @GetMapping("/rssi")
     public List<RSSID> getAll(){
         return rssiMapper.findAll();
     }
 
     @PostMapping("/rssi")
-    public String postMethod(@RequestParam(name = "pos_x") int pos_x, @RequestParam(name = "pos_y") int pos_y, @RequestBody List<RSSID> rssids){
+    public String postMethod(@RequestParam(name = "pos_x") float pos_x, @RequestParam(name = "pos_y") float pos_y, @RequestBody List<RSSID> rssids){
         for(int i=0; i<rssids.size(); i++){
             rssids.get(i).setPos_x(pos_x);
             rssids.get(i).setPos_y(pos_y);
